refactor(nuxt3): `defineNuxtLink` > `nuxt-link`, and some minors
