+++ packages/nuxt3/src/app/composables/index.ts
@@ -11,5 +11,5 @@ export type { CookieOptions, CookieRef } from './cookie'
 export { useRequestHeaders } from './ssr'
 export { abortNavigation, addRouteMiddleware, defineNuxtRouteMiddleware, navigateTo, useRoute, useRouter } from './router'
 export type { AddRouteMiddlewareOptions, RouteMiddleware } from './router'
-export { defineNuxtLink } from './defineNuxtLink'
-export type { DefineNuxtLinkOptions, NuxtLinkProps } from './defineNuxtLink'
+export { defineNuxtLink } from './nuxt-link'
+export type { DefineNuxtLinkOptions, NuxtLinkProps } from './nuxt-link'
